Day05/Study Assistent for Quiz Question Generation/requirements.txt
langchain>=0.0.350
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pydantic>=2.0.0
streamlit>=1.32.0 